docs: add audit reports for MultiSigControl contract
Add the Smart Contract audit reports conducted for the

Includes the audits for both the  MultiSigControl and the version containing the decentralised limits and controls.

These limits are expected to be used in early deployments of Vega and are designed to be raised and/or removed over time via the governance protocol as the security and the robustness of the implementation is demonstrated.